# Reduce memory usage and initialization time by 40% by using maps rather than plain objects