--- a/lib/add.js
+++ b/lib/add.js
@@ -4,11 +4,13 @@ var push = require('./util/add.js')
 
 module.exports = add
 
+var NO_CODES = []
+
 // Add `value` to the checker.
 function add(value, model) {
   var self = this
   var dict = self.data
-  var codes = dict[model]
+  var codes = dict[model] || NO_CODES
 
   push(dict, value, codes, self)