--- a/lib/add.js
+++ b/lib/add.js
@@ -10,7 +10,7 @@ var NO_CODES = []
 function add(value, model) {
   var self = this
   var dict = self.data
-  var codes = dict[model] || NO_CODES
+  var codes = dict.get(model) || NO_CODES
 
   push(dict, value, codes, self)