--- a/lib/index.js
+++ b/lib/index.js
@@ -49,7 +49,7 @@ function NSpell(aff, dic) {
 
   aff = affix(aff)
 
-  this.data = Object.create(null)
+  this.data = new Map()
   this.compoundRuleCodes = aff.compoundRuleCodes
   this.replacementTable = aff.replacementTable
   this.conversion = aff.conversion