--- a/lib/personal.js
+++ b/lib/personal.js
@@ -2,6 +2,8 @@
 
 module.exports = add
 
+var NO_CODES = []
+
 // Add a dictionary.
 function add(buf) {
   var self = this
@@ -36,7 +38,7 @@ function add(buf) {
       word = word.slice(1)
     }
 
-    self.add(word, model)
+    self.add(word, model, NO_CODES)
 
     if (forbidden) {
       self.data[word].push(flag)