--- a/lib/personal.js
+++ b/lib/personal.js
@@ -41,7 +41,7 @@ function add(buf) {
     self.add(word, model, NO_CODES)
 
     if (forbidden) {
-      self.data[word].push(flag)
+      self.data.get(word).push(flag)
     }
   }