--- a/lib/personal.js
+++ b/lib/personal.js
@@ -2,8 +2,6 @@
 
 module.exports = add
 
-var NO_CODES = []
-
 // Add a dictionary.
 function add(buf) {
   var self = this
@@ -38,7 +36,7 @@ function add(buf) {
       word = word.slice(1)
     }
 
-    self.add(word, model, NO_CODES)
+    self.add(word, model)
 
     if (forbidden) {
       self.data.get(word).push(flag)