--- a/lib/remove.js
+++ b/lib/remove.js
@@ -6,7 +6,7 @@ module.exports = remove
 function remove(value) {
   var self = this
 
-  delete self.data[value]
+  self.data.delete(value)
 
   return self
 }