--- a/lib/spell.js
+++ b/lib/spell.js
@@ -16,7 +16,7 @@ function spell(word) {
   // (whether `word` was compound).
   return {
     correct: self.correct(word),
-    forbidden: Boolean(value && flag(flags, 'FORBIDDENWORD', dict[value])),
-    warn: Boolean(value && flag(flags, 'WARN', dict[value]))
+    forbidden: Boolean(value && flag(flags, 'FORBIDDENWORD', dict.get(value))),
+    warn: Boolean(value && flag(flags, 'WARN', dict.get(value)))
   }
 }