--- a/lib/suggest.js
+++ b/lib/suggest.js
@@ -336,7 +336,7 @@ function generate(context, memory, words, edits) {
       result.push(value)
 
       corrected = form(context, value)
-      state = corrected && !flag(flags, noSuggestType, data[corrected])
+      state = corrected && !flag(flags, noSuggestType, data.get(corrected))
 
       memory.state[value] = state