--- a/lib/util/add.js
+++ b/lib/util/add.js
@@ -48,8 +48,6 @@ function add(dict, word, codes, options) {
     addRules(dict, word, codes)
   }
 
-  if (!codes) return
-
   position = -1
   length = codes.length