--- a/lib/util/add.js
+++ b/lib/util/add.js
@@ -12,15 +12,15 @@ var NO_RULES = []
 function addRules(dict, word, rules) {
   // Some dictionaries will list the same word multiple times with different
   // rule sets.
-  if (word in dict) {
-    var curr = dict[word]
+  if (dict.has(word)) {
+    var curr = dict.get(word)
     if (curr === NO_RULES) {
-      dict[word] = rules.concat()
+      dict.set(word, rules.concat())
     } else {
       push.apply(curr, rules)
     }
   } else {
-    dict[word] = rules.concat()
+    dict.set(word, rules.concat())
   }
 }
 
@@ -67,7 +67,7 @@ function add(dict, word, codes, options) {
       while (++offset < wordCount) {
         newWord = newWords[offset]
 
-        if (!(newWord in dict)) dict[newWord] = NO_RULES
+        if (!dict.has(newWord)) dict.set(newWord, NO_RULES)
 
         if (!rule.combineable) {
           continue
@@ -92,7 +92,7 @@ function add(dict, word, codes, options) {
 
           while (++suboffset < newWordCount) {
             otherNewWord = otherNewWords[suboffset]
-            if (!(otherNewWord in dict)) dict[otherNewWord] = NO_RULES
+            if (!dict.has(otherNewWord)) dict.set(otherNewWord, NO_RULES)
           }
         }
       }