--- a/lib/util/exact.js
+++ b/lib/util/exact.js
@@ -8,7 +8,7 @@ module.exports = exact
 function exact(context, value) {
   var data = context.data
   var flags = context.flags
-  var codes = data[value]
+  var codes = data.get(value)
   var compound
   var index
   var length