--- a/lib/util/form.js
+++ b/lib/util/form.js
@@ -21,7 +21,7 @@ function form(context, value, all) {
   value = normalize(value, context.conversion.in)
 
   if (exact(context, value)) {
-    if (!all && flag(flags, 'FORBIDDENWORD', dict[value])) {
+    if (!all && flag(flags, 'FORBIDDENWORD', dict.get(value))) {
       return null
     }
 
@@ -32,7 +32,7 @@ function form(context, value, all) {
   if (value.toUpperCase() === value) {
     alternative = value.charAt(0) + value.slice(1).toLowerCase()
 
-    if (ignore(flags, dict[alternative], all)) {
+    if (ignore(flags, dict.get(alternative), all)) {
       return null
     }
 
@@ -45,7 +45,7 @@ function form(context, value, all) {
   alternative = value.toLowerCase()
 
   if (alternative !== value) {
-    if (ignore(flags, dict[alternative], all)) {
+    if (ignore(flags, dict.get(alternative), all)) {
       return null
     }