Benchmarks performed by initializing the french dictionary, on node v14 and using the version of nspell that includes the previous performance optimizations I submitted.

## Before

CPU usage:

![image](https://user-images.githubusercontent.com/1812093/95529858-0e25b100-09d4-11eb-8712-7386a1a1876f.png)

RAM usage:

![image](https://user-images.githubusercontent.com/1812093/95529933-4c22d500-09d4-11eb-835b-dc194ed2ed16.png)

## After

CPU usage:

![image](https://user-images.githubusercontent.com/1812093/95529946-580e9700-09d4-11eb-8f17-8e798aeb7d3e.png)

RAM usage:

![image](https://user-images.githubusercontent.com/1812093/95529957-60ff6880-09d4-11eb-9c91-a6e05634665a.png)

## Summary

- Pros:
  - CPU usage went down by a further ~40%, now on my machine nspell takes ~800ms to run the regexes, ~800ms to check for the existence of the potential new words in the dictionary, ~1.5s to add new keys to the dictionary map, and another ~200ms doing other things. There's basically near zero room for further optimizations on our end I think, unless using some other data structure for some reason makes the engine run the code faster.
  - RAM usage went down by ~40% too, which for the french dictionary that means ~160MB less RAM needed, which I think is pretty significant. Also while running the profiler the peak memory used went from ~600MB to ~250MB, but I'm not sure how reliable that counter is.
- Cons:
  - A Map now is used rather than a plain object for storing the dictionary, which is what provided the performance gains in this PR.
  - As the `data` property is meant to be exposed making it a Map makes this a potentially breaking change.
  - A native Map implementation must be available for these gains to materialize, otherwise we'd obviously still be dealing with plain objects under the hood.
  - A somewhat recent version of v8 is needed, as running the benchmark on node v12.12 (v8 v7.something) made this actually significantly slower than before (~8s rather than the previous ~6s), but running it on node 14 (v8 v8.something) made this run at the reported speeds.

The cons fortunately are not relevant for me, but if supporting IE10 or something is more important for nspell than the performance gains feel free not to merge this PR. IMHO publishing this as v3 would be the best option.