@wooorm 2020-10-09
Nice! It’s indeed a breaking change. I’m not too worried about old IE, as most folks will do this in Node (or Electron or so?). I am more worried about this being significantly slower on an LTS Node thoug. I don’t think we should wait till Node 12 is end of life (2022-04-30), but...

Do you know if V8 gets updated in LTS releases? From what I see it looks like they’re cherry picking a couple things but otherwise not updating, right?