@fabiospampinato 2020-10-09
> Do you know if V8 gets updated in LTS releases? From what I see it looks like they’re cherry picking a couple things but otherwise not updating, right?

I don't really know anything about that, but [this](https://nodejs.org/en/download/releases/) lists all Node releases along with the used v8 version, I did some tests and on Node v13.14 which use v8 v7.9.317.25 this is still fast, on Node v13.0 which uses v8 v7.8.279.17 this is slow, so the magic happened at some point between v7.8 and v7.9, and also according to that table during Node v12 iteration cycle they updated v8 from v7.7 to v7.8, so I guess it's possible that they may eventually update it to v7.9 too, no idea if that will actually happen or what the timeline for that would be though.

Potentially we could make the library throw an error if the version of v8 is not high enough, telling users to upgrade Node or downgrade nspell 🤷‍♂️