@wooorm 2020-10-09
> Potentially we could make the library throw an error if the version of v8 is not high enough, telling users to upgrade Node or downgrade nspell 🤷‍♂️

Node 12 is recommended for most users. I don’t think throwing an error for most users is feasible.

---

An alternative is to use an object based class normally (if this is still fast):

```js
function MiniMap() {
  this.data = Object.create(null)
}

MiniMap.prototype.get = get
MiniMap.prototype.has = has
MiniMap.prototype.delete = remove

function get(key) {
  return this.data[key]
}
function set(key, value) {
  this.data[key] = value
}
function has(key) {
  return key in this.data
}
function remove(key) {
  delete this.data[value]
}
```

...then, upon creation do:

```js
var Dictionary = UserGivenMap ? new UserGivenClass() : new MiniMap()
```

And document that folks can pass a `Map` in on Node 14 for better performance?
If this has roughly the same performance as the current object based solution, it would not be a breaking feature, and allow folks on new V8s to opt-in to a performance improvement.