@fabiospampinato 2020-10-09
> Node 12 is recommended for most users. I don’t think throwing an error for most users is feasible.

It's not as terrible as it sounds IMHO, error -> downgrade -> done.

> An alternative is to use an object based class normally (if this is still fast):
> And document that folks can pass a Map in on Node 14 for better performance?

It's an option, but that would probably mean that most current users may never learn about this potential optimization, and eventually as all future supported Node versions will ship with a fast Map implementation that would become a kind of a useless option, plus isn't this still a breaking change since the interface of the `data` property would change anyway?

I don't know, which way do you want to take?