@brianhatchet 2021-03-12
https://nodejs.org/en/about/releases/

Node JS's LTS is v14 now.

https://nodejs.org/en/blog/release/v14.0.0/

They've upgraded to v8 version 8.1 in that release. So this should be pretty safe.