@wooorm 2021-03-13
Yup! I also started switching my projects to ESM only because the last Node w/o them out of the box (10) is EOL soon.

I am a bit worried about non-V8 though: if this was super slow in any non-V8 engine, that’d suck.

But using maps soon sounds good to me