@fabiospampinato 2021-03-13
@wooorm IMHO non-v8s are largely not an issue, there are basically only ~20% of reasonable non-v8s out there (meaning either Firefox or Safari, the rest is either Chrome, other flavors of Chromium or junk like IE), if Firefox or Safari run this significantly slower just because it's using a Map under the hood (I doubt it) it's kind of their problems and in any case downgrading for users is always an option. 

At the end of the day though as outlined in #11 this just can't be made really fast if one has to parse dictionaries at runtime, we need to switch to something fundamentally better, I think @brianhatchet, with whom I chatted the other day, would first like to help close the gap with hunspell with regards to supported affix rules, from that perspective I think this PR is largely irrelevant, it has nothing to do with compounding rules and performance-wise it'll become irrelevant once the switch is made to the system that can work with pre-parsed dictionaries.