@wooorm 2021-03-13
> Besides are people even doing spell-checking on the server? That sounds weird to me 🤔

You mean the client?  In any case, folks are doing spell-checking everywhere, yes.

> IMHO non-v8s are largely not an issue, there are basically only ~20% of reasonable non-v8s

If this library is super slow for 1/5 people, then that’s a bug.
In any case, this has to be weighed of course, depending on how much the win is for the other 80%.
But I’m not sure what your 20% refers to.

> once the switch is made to the system that can work with pre-parsed dictionaries.

I do not know what this entails, and what your conversation was. Are you re-inventing hunspell dictionaries? That sounds like something mostly unrelated to nspell?